requirements.txt
browser-use
python-dotenv
openai
fastapi
uvicorn
playwright
asyncio